+++ src/lib/components/Machine/PointCloudObject.svelte
@@ -226,7 +226,6 @@
 							...hierarchy.parentTraits(geometriesInFrame.referenceFrame),
 							traits.Matrix(center.toMatrix4()),
 							traits.Geometry(geometry),
-							traits.Opacity(0.2),
 							traits.Color({ r: 0, g: 1, b: 0 }),
 							traits.PointCloudObjectAPI,
 						]