+++ src/lib/plugins/Monitor/__tests__/MonitorDetails.svelte.spec.ts
@@ -39,7 +39,10 @@ describe('MonitorDetails', () => {
 		})
 		vi.mocked(useConfigFrames.useConfigFrames).mockReturnValue({
 			unsetFrames: [],
+			unresolvedFrames: new Set(),
 			current: {},
+			fragmentFrames: {},
+			effectiveFrames: new Map(),
 		})
 		vi.mocked(useLinkedEntities.useLinkedEntities).mockReturnValue({
 			current: [],